requirements.txt
python 3.9.1
django 3.0
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
